handle failures in continuous flashing mode
when flashing failed, e.g. we aborted if we couldn't upload the routeros
npk, the device will send a ton of retry and finally a few terminate
commands to us. the sleep(10) meant that we were essentially blocked
from restarting the loop, until we (slowly) read all those back. since
read() is blocking anyways, just remove the unconditional sleep.