verify firmware image before attempting installation
previously, we formatted the flash first, asked questions later. this
could have* left devices without an operating system.

note that this asks the plugin twice (and for URLs also asks the web
server twice); this theoretically introduces a TOCTOU, but for static
assets like these it shouldn't matter in practice.

* read: "did"